Modified List class to use  ArrayList<Task> to store tasks instead.
