--- src/main/java/List.java
@@ -1,37 +1,0 @@
-public class List {
-
-    private Task[] tasks;
-    private int index = 0;
-
-    public List(int max) {
-        tasks = new Task[max];
-    }
-
-    public String addToList(Task newTask) {
-        tasks[index] = newTask;
-        index++;
-        return newTask.toString();
-    }
-
-    public String taskDone(int id) {
-        tasks[id - 1].markAsDone();
-        return tasks[id - 1].toString();
-    }
-
-    public int taskCount() {
-        return index;
-    }
-
-    public String getList() {
-        int counter = 1;
-        String result = "";
-
-        for (int i = 0; i < tasks.length; i++) {
-            if (tasks[i] != null) {
-                result = result + counter + "." + tasks[i] + "\n";
-                counter++;
-            }
-        }
-        return result;
-    }
-}
